ODP-5751: Clean up Python code
