+++ knox-enablement/.gitignore
@@ -1,0 +1,3 @@
+venv
+build
+knox_setup.egg-info
