+++ knox-enablement/knox_setup/main.py
@@ -2,15 +2,15 @@
 """
 Knox Enablement for Ambari BigData Cluster
 """
-
+import argparse
 import sys
-from steps import (
-    set_proxy_users, 
-    start_knox, 
-    update_topology, 
-    update_whitelist, 
-    restart_services, 
-    export_knox_cert, 
+from knox_setup.steps import (
+    set_proxy_users,
+    start_knox,
+    update_topology,
+    update_whitelist,
+    restart_services,
+    export_knox_cert,
     ambari_sso_setup,
     ambari_ldap_setup,
     import_knox_cert,
@@ -69,7 +69,7 @@ FLOWS = {
 def run_flow(flow_name: str):
     """Execute a named flow (sequence of steps)."""
     steps = FLOWS[flow_name]
-    
+
     print("=" * 60)
     print(f"KNOX ENABLEMENT - {flow_name.upper()}")
     print(f"Steps: {' → '.join(steps)}")
@@ -86,35 +86,40 @@ def run_flow(flow_name: str):
     print("=" * 60)
 
 
-def print_usage():
-    """Print usage information."""
-    print("Usage: python main.py <step_or_flow>")
-    print("\nAvailable steps:")
-    for step_name in STEPS:
-        print(f"  - {step_name}")
-    print("\nAvailable flows:")
-    for flow_name, steps in FLOWS.items():
-        print(f"  - {flow_name} ({' → '.join(steps)})")
 
 
 def main():
     """Main entry point."""
-    if len(sys.argv) < 2:
-        print_usage()
-        sys.exit(1)
 
-    target = sys.argv[1]
+    all_arg_names = list(STEPS.keys() | FLOWS.keys())
+
+    epilogue_text = "\n\t- ".join(["Valid targets are: "]+sorted(all_arg_names))
+
+    parser = argparse.ArgumentParser(description="Knox Enablement for Ambari", epilog=epilogue_text, formatter_class=argparse.RawTextHelpFormatter)
+    parser.add_argument("--with-https", "-s", action="store_true", help="Use SSL for Ambari URLs (default: http)")
+    parser.add_argument("--with-knox-http", "-k", action="store_false", help="Disable using https for Ambari SSO. (default: https)")
+    parser.add_argument("target", metavar="TARGET", choices=all_arg_names, help="Step or flow to execute")
+
+    args = parser.parse_args()
+
+    protocol = "https" if args.with_https else "http"
+    knox_protocol = "http" if args.with_knox_http else "https"
+
+    target = args.target
 
     if target in FLOWS:
         run_flow(target)
     elif target in STEPS:
         print(f"Executing: {target}")
-        STEPS[target]()
+        if target == "ambari_sso_setup":
+            STEPS[target](protocol=knox_protocol)
+        else:
+            STEPS[target](protocol)
         print(f"\nCompleted: {target}")
     else:
         print(f"Unknown step or flow: {target}")
-        print_usage()
-        sys.exit(1)
+        print_usage(sys.argv[0])
+        return 1
 
 
 if __name__ == "__main__":
