+++ knox-enablement/knox_setup/modules/ambari_state.py
@@ -4,12 +4,11 @@ Fetches values once from API and caches them for reuse.
 """
 
 import urllib.request
-import urllib.error
 import ssl
 import json
 import base64
 from typing import Optional
-from config.settings import AmbariConfig
+from knox_setup.config.settings import AmbariConfig
 
 
 class AmbariState:
@@ -17,45 +16,45 @@ class AmbariState:
     Singleton-like class to cache Ambari cluster state.
     Fetches values from API once and stores them globally.
     """
-    
+
     _instance: Optional['AmbariState'] = None
     _initialized: bool = False
-    
+
     # Cached values
     _cluster_name: Optional[str] = None
     _hosts: Optional[list] = None
     _services: Optional[list] = None
-    
+
     def __new__(cls):
         if cls._instance is None:
             cls._instance = super().__new__(cls)
         return cls._instance
-    
+
     def __init__(self):
         if AmbariState._initialized:
             return
         AmbariState._initialized = True
-    
+
     def _api_request(self, endpoint: str) -> dict:
         """Make GET request to Ambari API."""
         url = f"{AmbariConfig.PROTOCOL}://{AmbariConfig.HOST}:{AmbariConfig.PORT}{endpoint}"
-        
+
         auth = base64.encodebytes(
             f'{AmbariConfig.USERNAME}:{AmbariConfig.PASSWORD}'.encode()
         ).decode().replace('\n', '')
-        
+
         request = urllib.request.Request(url)
         request.add_header('Authorization', f'Basic {auth}')
         request.add_header('X-Requested-By', 'ambari')
-        
+
         ctx = ssl.create_default_context()
         ctx.check_hostname = False
         ctx.verify_mode = ssl.CERT_NONE
-        
+
         response = urllib.request.urlopen(request, context=ctx)
         body = response.read().decode('utf-8')
         return json.loads(body) if body else {}
-    
+
     @property
     def cluster_name(self) -> str:
         """Get cluster name (fetched once from API)."""
@@ -68,8 +67,9 @@ class AmbariState:
             # Use the first cluster
             self._cluster_name = items[0]["Clusters"]["cluster_name"]
             print(f"[AmbariState] Cluster name: {self._cluster_name}")
+        assert(self._cluster_name is not None)
         return self._cluster_name
-    
+
     @property
     def hosts(self) -> list:
         """Get list of hosts in the cluster (fetched once from API)."""
@@ -78,12 +78,12 @@ class AmbariState:
             endpoint = f"/api/v1/clusters/{self.cluster_name}/hosts"
             result = self._api_request(endpoint)
             self._hosts = [
-                item["Hosts"]["host_name"] 
+                item["Hosts"]["host_name"]
                 for item in result.get("items", [])
             ]
             print(f"[AmbariState] Found {len(self._hosts)} hosts")
         return self._hosts
-    
+
     @property
     def first_host(self) -> str:
         """Get the first host in the cluster."""
@@ -91,7 +91,7 @@ class AmbariState:
         if not hosts:
             raise RuntimeError("No hosts found in cluster")
         return hosts[0]
-    
+
     @property
     def services(self) -> list:
         """Get list of installed services (fetched once from API)."""
@@ -105,11 +105,11 @@ class AmbariState:
             ]
             print(f"[AmbariState] Found {len(self._services)} services")
         return self._services
-    
+
     def is_service_installed(self, service_name: str) -> bool:
         """Check if a service is installed."""
         return service_name in self.services
-    
+
     def refresh(self):
         """Clear cached values to force re-fetch on next access."""
         self._cluster_name = None
@@ -125,7 +125,7 @@ _state: Optional[AmbariState] = None
 def get_ambari_state(refresh: bool = False) -> AmbariState:
     """
     Get the global AmbariState instance.
-    
+
     Args:
         refresh: If True, clears cached values to force re-fetch from API
     """
