+++ knox-enablement/pyproject.toml
@@ -1,0 +1,98 @@
+[project]
+name = "knox-setup"
+description = "Automated Knox Setup for Ambari"
+readme = "README.md"
+version = "1.0.0"
+license = "Apache-2.0"
+requires-python = ">=3.11"
+dependencies = [
+  "paramiko>=3.4.0",
+  "requests>=2.31.0",
+  "python-dotenv>=1.0.0",
+  "urllib3>=2.0.0",
+]
+
+authors = [
+  { name = "Harshith Ghande", email = "harshith@acceldata.io" },
+  { name = "Jeffrey Smith", email = "jeffrey.smith@acceldata.io" },
+]
+maintainers = [{ name = "Jeffrey Smith", email = "jeffrey.smith@acceldata.io" }]
+keywords = ["knox", "setup", "automation", "ssh", "automated"]
+classifiers = [
+  "Intended Audience :: Developers",
+  "Programming Language :: Python :: 3.11",
+]
+
+[project.scripts]
+knox-setup = "knox_setup.main:main"
+
+[project.urls]
+Homepage = "https://github.com/acceldata-io/ce-utils"
+
+[tool.setuptools.packages.find]
+include = ["knox_setup*"]
+
+[tool.setuptools.package-data]
+"knox_setup" = ["templates/**/*"]
+
+[tool.ruff]
+# Exclude a variety of commonly ignored directories.
+exclude = [
+  ".bzr",
+  ".direnv",
+  ".eggs",
+  ".git",
+  ".git-rewrite",
+  ".hg",
+  ".ipynb_checkpoints",
+  ".mypy_cache",
+  ".nox",
+  ".pants.d",
+  ".pyenv",
+  ".pytest_cache",
+  ".pytype",
+  ".ruff_cache",
+  ".svn",
+  ".tox",
+  ".venv",
+  ".vscode",
+  "__pypackages__",
+  "_build",
+  "buck-out",
+  "build",
+  "dist",
+  "node_modules",
+  "site-packages",
+  "venv",
+]
+line-length = 120
+indent-width = 4
+target-version = "py311"
+
+[tool.ruff.lint]
+# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
+select = ["E4", "E7", "E9", "F"]
+ignore = []
+
+# Allow fix for all enabled rules (when `--fix`) is provided.
+fixable = ["ALL"]
+unfixable = []
+
+# Allow unused variables when underscore-prefixed.
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
+
+[tool.pip-licenses]
+fail-on = "GPL"
+format = "plain-vertical"
+from = "mixed"
+no-license-path = true
+output-file = "LICENSES"
+with-authors = true
+with-license-file = true
+with-urls = true
