--- knox-enablement/requirements.txt
@@ -1,5 +1,0 @@
-paramiko>=3.4.0
-requests>=2.31.0
-python-dotenv>=1.0.0
-urllib3>=2.0.0
-
