--- knox-enablement/steps/__init__.py
@@ -1,25 +1,0 @@
-from steps import set_proxy_users
-from steps import start_knox
-from steps import update_topology
-from steps import update_whitelist
-from steps import restart_services
-from steps import export_knox_cert
-from steps import ambari_sso_setup
-from steps import ambari_ldap_setup
-from steps import import_knox_cert
-from steps import restart_ambari
-from steps import add_ranger_policy
-
-__all__ = [
-    "set_proxy_users", 
-    "start_knox", 
-    "update_topology", 
-    "update_whitelist", 
-    "restart_services", 
-    "export_knox_cert", 
-    "ambari_sso_setup",
-    "ambari_ldap_setup",
-    "import_knox_cert",
-    "restart_ambari",
-    "add_ranger_policy",
-]
